24년 04월 02일 정다현

+++ DreamSwellProject/src/main/java/kr/dream/swell/dao/DreamBoardDAO.java
@@ -18,7 +18,7 @@ public interface DreamBoardDAO {
 	 */
 	ArrayList<DreamSwellBoardVO> selectScrollList(HashMap<String, Object> map) throws SQLException;
 	/** board의 최대 idx 리턴 */
-	int findLastItemIdx() throws SQLException;
+	Long findLastItemIdx() throws SQLException;
 	
 	DreamSwellBoardVO selectByIdx(Long idx) throws SQLException;
 	Long selectCount(HashMap<String, Object> map) throws SQLException;
